+++ resources/views/posts/show.blade.php
@@ -4,18 +4,22 @@
     <a style="margin-bottom: 10px;" href="/posts" class="btn btn-default">Go Back</a>
     @if (!empty($post))
     <div class="well">
+        <img style="width: 20%; height:20%" src="/storage/cover_image/{{$post->cover_image}}" alt="">
+        <br><br>
         <h3>{{$post->title}}</h3>
         <div>
             {{$post->body}}
         </div>
-        <small>Written on {{$post->created_at}}</small>
+        <small>Written by {{$post->user->name}} at {{$post->created_at}}</small>
     </div>
     <hr>
-    <a href="/posts/{{$post->id}}/edit" class="btn btn-default">Edit</a>
 
-    {!! Form::open(['action' => ['PostsController@destroy', $post->id], 'method'=> 'DELETE', 'class' => 'pull-right']) !!}
-        {{Form::submit('Delete', ['class' => 'btn btn-danger'])}}
-    {!! Form::close() !!}
+    @if (!Auth::guest() && Auth::user()->id == $post->user_id)
+        <a href="/posts/{{$post->id}}/edit" class="btn btn-default">Edit</a>
+        {!! Form::open(['action' => ['PostsController@destroy', $post->id], 'method'=> 'DELETE', 'class' => 'pull-right']) !!}
+            {{Form::submit('Delete', ['class' => 'btn btn-danger'])}}
+        {!! Form::close() !!}
+    @endif
     @else
         <p>No Post is available!</p>
     @endif
